+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- ESM module
+- Karaoke feature
+
+### Changed
+
+- **BREAKING CHANGE**: rename UMD module
+
+### Removed
+
+- **BREAKING CHANGE**: height option. You can set height through CSS.
+
 ## [1.3.0]
 
 ### Changed
@@ -96,7 +109,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Lyrics synchronizer
 - Three sizes: tiny, medium, full.
 
-[Unreleased]: https://github.com/guoyunhe/rabbit-lyrics/compare/v1.3.0...HEAD
+[unreleased]: https://github.com/guoyunhe/rabbit-lyrics/compare/v1.3.0...HEAD
 [1.3.0]: https://github.com/guoyunhe/rabbit-lyrics/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/guoyunhe/rabbit-lyrics/compare/v1.1.2...v1.2.0
 [1.1.2]: https://github.com/guoyunhe/rabbit-lyrics/compare/v1.1.1...v1.1.2